Multiprocessing request
