+++ Database/database.c
@@ -1,212 +1,49 @@
+#include <fcntl.h>
+#include <sys/stat.h>
+#include <sys/types.h>
+#include <sys/time.h>
+#include <unistd.h>
 #include <stdio.h>
-#include <string.h>
-#include <sqlite3.h>
-#include "database.h"
+#include <time.h>
+#include "../Api/commons.h"
+#include "../Api/namedPipe/namedPipe.h"
+#include "databaseapi.h"
 
-int printRow(void *NotUsed, int argc, char **argv, char **azColName) {
-  NotUsed = 0;
-  for (int i = 0; i < argc; i=i+2) {
-      printf("%s\t%s", argv[i] ? argv[i] : "NULL", argv[i+1] ? argv[i+1] : "NULL");
-  }
-  printf("\n");
-  return 0;
-}
 
-sqlite3* DbOpen(void) {
-  sqlite3* db;
-  int rc = sqlite3_open(DATABASE_NAME, &db);
-  if (rc != SQLITE_OK) {
-    fprintf(stderr, "Cannot open database: %s\n", sqlite3_errmsg(db));
-    sqlite3_close(db);
-    return NO_SESSION;
-  }
-  return db;
-}
+int main (int argc, char const *argv[]){
 
-int DbCreateTable() {
-  sqlite3* db;
-  int rc = sqlite3_open(DATABASE_NAME, &db);
+	fd_set set;  	//fds to monitor
+	Request * r;
+	struct timeval tv;
+	Connection* c;
 
-  if (rc != SQLITE_OK) {
-    fprintf(stderr, "Cannot open database: %s\n", sqlite3_errmsg(db));
-    sqlite3_close(db);
-    return -1;
-  }
-
-  char *err_msg = 0;
-  int req = sqlite3_exec(db, SQL_CREATE_TABLE, 0, 0, &err_msg);
-  if (req != SQLITE_OK ) {
-    fprintf(stderr, "SQL error: %s\n", err_msg);
-    fprintf(stderr, "SQL error: %d\n", req);
-    sqlite3_free(err_msg);
-    return -1;
-  }
-  return req;
-}
-
-int DbCeckTableExistance(sqlite3* db) {
-  char *err_msg = 0;
-  int rc = sqlite3_exec(db, SQL_SELECT_ALL, 0, 0, &err_msg);
-
-  if (rc != SQLITE_OK ) {
-    fprintf(stderr, "SQL error: %s\n", err_msg);
-    fprintf(stderr, "SQL error: %d\n", rc);
-
-    if ( rc == -1 ) {
-      sqlite3_close(db);
-    }
-  }
-  return rc;
-}
-
-int DbAddStudent(char name[25], char average[5]) {
-
-  int rc;
-  sqlite3* db = DbOpen();
-  rc = DbCeckTableExistance(db);
-
-  if ( db == NULL || rc == -1 ) {
-    sqlite3_close(db);
-    return DATABASE_ERROR;
-  }
-
-	char *err_msg = 0;
-
-  char sql[200];
-  sprintf (sql, SQL_ADD_STUDENT, name, average);
-  printf("%s\n", sql);
-
-	rc = sqlite3_exec(db, sql, 0, 0, &err_msg);
-
-	if (rc != SQLITE_OK ) {
-    if ( rc == USER_EXISTS ) {
-      printf("%s\n", USER_EXISTS_ERROR_MESSAGE);
-      return ADD_STUDENT_USER_EXIST;
-    } else {
-      fprintf(stderr, "SQL error: %s\n", err_msg);
-      fprintf(stderr, "SQL error: %d\n", rc);
-      return DATABASE_ERROR;
-    }
-		sqlite3_free(err_msg);
-		sqlite3_close(db);
+	// //initialize request queue TODO REDO
+	if( (c = openConnection(SQL_QUEUE)) == NOT_FOUND) {
+		return ERROR_OPEN_REQUEST_QUEUE;
 	}
-	sqlite3_close(db);
-	return ADD_STUDENT_SUCCESS;
+	//sets up the fds to monitor
+	monitorConnection(c, &set);
+
+	//waits 2 seconds
+	tv.tv_sec = 2;
+  	tv.tv_usec = 0;
+
+	while(1) {
+		int fdCount = select(c -> fd + 1, &set, NULL, NULL, &tv);
+		tv.tv_sec = 2;
+		monitorConnection(c, &set);
+		if( fdCount > 0){
+			r = getRequest(c);
+			if( r != NOT_FOUND ){
+				printf("Request procesada\n");
+				processRequestDatabase (r);
+			}
+		} else if ( fdCount < 0 ){
+			printf("Error\n");
+		} else {
+			printf("No requests to process\n");
+		}
+	}
+
+	return 0;
 }
-
-int DbReadStudents (){
-  int rc;
-  sqlite3* db = DbOpen();
-  rc = DbCeckTableExistance(db);
-
-  if ( db == NULL || rc == -1 ) {
-    sqlite3_close(db);
-    return DATABASE_ERROR;
-  }
-
-  char *err_msg = 0;
-
-  printf("Name\tAverage\n");
-  rc = sqlite3_exec(db, SQL_SELECT_ALL  , printRow, 0, &err_msg);
-
-  if (rc != SQLITE_OK ) {
-    fprintf(stderr, "Failed to select data\n");
-    fprintf(stderr, "SQL error: %s\n", err_msg);
-
-    sqlite3_free(err_msg);
-    sqlite3_close(db);
-    
-    return 1;
-  } 
-  
-  sqlite3_close(db);
-  return READ_STUDENTS_SUCCESS;
-}
-
-int DbDeleteStudent (char name[25]){
-
-  int rc;
-  sqlite3* db = DbOpen();
-  rc = DbCeckTableExistance(db);
-
-  if ( db == NULL || rc == -1 ) {
-    sqlite3_close(db);
-    return DATABASE_ERROR;
-  }
-
-  char *err_msg = 0;
-
-
-  char sql[200];
-  sprintf (sql, SQL_DELETE_STUDENT, name);
-  printf("%s\n", sql);
-  rc = sqlite3_exec(db, sql, 0, 0, &err_msg);
-
-  if (rc != SQLITE_OK ) {
-    fprintf(stderr, "SQL error: %s\n", err_msg);
-    fprintf(stderr, "SQL error: %d\n", rc);
-    sqlite3_free(err_msg);
-    sqlite3_close(db);
-  }
-
-  sqlite3_close(db);
-  return DELETE_STUDENT_SUCCESS;
-}
-
-int DbUpdateStudent (char currentName[25], char newName[25], char average[5]) {
-  int rc;
-  sqlite3* db = DbOpen();
-  rc = DbCeckTableExistance(db);
-
-  if ( db == NULL || rc == -1 ) {
-    sqlite3_close(db);
-    return DATABASE_ERROR;
-  }
-
-  char *err_msg = 0;
-
-  char sql[200];
-  sprintf (sql, SQL_UPDATE_STUDENT, newName, average, currentName);
-  printf("%s\n", sql);
-  
-  rc = sqlite3_exec(db, sql, 0, 0, &err_msg);
-
-  if (rc != SQLITE_OK ) {
-    fprintf(stderr, "SQL error: %s\n", err_msg);
-    fprintf(stderr, "SQL error: %d\n", rc);
-    sqlite3_free(err_msg);
-    sqlite3_close(db);
-  }
-
-  sqlite3_close(db);
-  return UPDATE_STUDENT_SUCCESS;
-}
-
-int DbDropTable () {
-
-  int rc;
-  sqlite3* db = DbOpen();
-  rc = DbCeckTableExistance(db);
-
-  if ( db == NULL || rc == -1 ) {
-    sqlite3_close(db);
-    return DATABASE_ERROR;
-  }
-  char *err_msg = 0;
-
-  rc = sqlite3_exec(db, SQL_DROP_TABLE, 0, 0, &err_msg);
-
-  if (rc != SQLITE_OK ) {
-    fprintf(stderr, "SQL error: %s\n", err_msg);
-    fprintf(stderr, "SQL error: %d\n", rc);
-    sqlite3_free(err_msg);
-    sqlite3_close(db);
-  }
-
-  sqlite3_close(db);
-  return 0;
-}
-
-
-
-
