+++ Server/clientServer.c
@@ -5,6 +5,7 @@
 #include <unistd.h>
 #include <stdio.h>
 #include <time.h>
+#include "../Api/namedPipe/namedPipe.h"
 #include "../Api/commons.h"
 
 int main (int argc, char const *argv[]){
@@ -15,7 +16,7 @@ int main (int argc, char const *argv[]){
 	Connection* c;
 
 	// //initialize request queue TODO REDO
-	if( (c = openConnection()) == NOT_FOUND) {
+	if( (c = openConnection(REQUEST_QUEUE)) == NOT_FOUND) {
 		return ERROR_OPEN_REQUEST_QUEUE;
 	}
 	//sets up the fds to monitor
@@ -26,6 +27,8 @@ int main (int argc, char const *argv[]){
   	tv.tv_usec = 0;
 
 	while(1) {
+		int pid = getpid();
+		printf("pid ->%d -<", pid);
 		int fdCount = select(c -> fd + 1, &set, NULL, NULL, &tv);
 		tv.tv_sec = 2;
 		monitorConnection(c, &set);
@@ -33,7 +36,7 @@ int main (int argc, char const *argv[]){
 			r = getRequest(c);
 			if( r != NOT_FOUND ){
 				printf("Request procesada\n");
-				processRequest(r);
+				processRequestServer(r);
 			}
 		} else if ( fdCount < 0 ){
 			printf("Error\n");
